refactor(datagrid-web): move check to module
